+++ pennylane_lightning/src/algorithms/Observables.cpp
@@ -1,4 +1,4 @@
-// Copyright 2021 Xanadu Quantum Technologies Inc.
+// Copyright 2022 Xanadu Quantum Technologies Inc.
 
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,10 +12,16 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "JacobianTape.hpp"
+#include "Observables.hpp"
 
-template class Pennylane::Algorithms::OpsData<float>;
-template class Pennylane::Algorithms::OpsData<double>;
+template class Pennylane::Algorithms::NamedObs<float>;
+template class Pennylane::Algorithms::NamedObs<double>;
 
-template class Pennylane::Algorithms::JacobianData<float>;
-template class Pennylane::Algorithms::JacobianData<double>;
+template class Pennylane::Algorithms::HermitianObs<float>;
+template class Pennylane::Algorithms::HermitianObs<double>;
+
+template class Pennylane::Algorithms::TensorProdObs<float>;
+template class Pennylane::Algorithms::TensorProdObs<double>;
+
+template class Pennylane::Algorithms::Hamiltonian<float>;
+template class Pennylane::Algorithms::Hamiltonian<double>;