+++ pennylane_lightning/src/bindings/Bindings.cpp
@@ -19,12 +19,15 @@
 
 #include "GateUtil.hpp"
 #include "SelectKernel.hpp"
+#include "StateVecAdjDiff.hpp"
 #include "StateVectorManagedCPU.hpp"
 
 #include "pybind11/pybind11.h"
 
 /// @cond DEV
 namespace {
+using namespace Pennylane;
+using namespace Pennylane::Util;
 using namespace Pennylane::Algorithms;
 using namespace Pennylane::Gates;
 
@@ -53,8 +56,6 @@ void lightning_class_bindings(py::module_ &m) {
 
     using np_arr_c = py::array_t<std::complex<ParamT>,
                                  py::array::c_style | py::array::forcecast>;
-    using np_arr_r =
-        py::array_t<ParamT, py::array::c_style | py::array::forcecast>;
     using sparse_index_type =
         long int; // Kokkos Kernels needs signed int as Ordinal type.
     using np_arr_sparse_ind =
@@ -75,246 +76,6 @@ void lightning_class_bindings(py::module_ &m) {
 
     pyclass.def("kernel_map", &svKernelMap<PrecisionT>,
                 "Get internal kernels for operations");
-
-    //***********************************************************************//
-    //                              Observable
-    //***********************************************************************//
-
-    class_name = "ObsStructC" + bitsize;
-    using obs_data_var = std::variant<std::monostate, np_arr_r, np_arr_c>;
-    py::class_<ObsDatum<PrecisionT>>(m, class_name.c_str(), py::module_local())
-        .def(py::init([](const std::vector<std::string> &names,
-                         const std::vector<obs_data_var> &params,
-                         const std::vector<std::vector<size_t>> &wires) {
-            std::vector<typename ObsDatum<PrecisionT>::param_var_t> conv_params(
-                params.size());
-            for (size_t p_idx = 0; p_idx < params.size(); p_idx++) {
-                std::visit(
-                    [&](const auto &param) {
-                        using p_t = std::decay_t<decltype(param)>;
-                        if constexpr (std::is_same_v<p_t, np_arr_c>) {
-                            auto buffer = param.request();
-                            auto ptr =
-                                static_cast<std::complex<ParamT> *>(buffer.ptr);
-                            if (buffer.size) {
-                                conv_params[p_idx] =
-                                    std::vector<std::complex<ParamT>>{
-                                        ptr, ptr + buffer.size};
-                            }
-                        } else if constexpr (std::is_same_v<p_t, np_arr_r>) {
-                            auto buffer = param.request();
-
-                            auto *ptr = static_cast<ParamT *>(buffer.ptr);
-                            if (buffer.size) {
-                                conv_params[p_idx] =
-                                    std::vector<ParamT>{ptr, ptr + buffer.size};
-                            }
-                        } else {
-                            PL_ABORT(
-                                "Parameter datatype not current supported");
-                        }
-                    },
-                    params[p_idx]);
-            }
-            return ObsDatum<PrecisionT>(names, conv_params, wires);
-        }))
-        .def("__repr__",
-             [](const ObsDatum<PrecisionT> &obs) {
-                 using namespace Pennylane::Util;
-                 std::ostringstream obs_stream;
-                 std::string obs_name = obs.getObsName()[0];
-                 for (size_t o = 1; o < obs.getObsName().size(); o++) {
-                     if (o < obs.getObsName().size()) {
-                         obs_name += " @ ";
-                     }
-                     obs_name += obs.getObsName()[o];
-                 }
-                 obs_stream << "'wires' : " << obs.getObsWires();
-                 return "Observable: { 'name' : " + obs_name + ", " +
-                        obs_stream.str() + " }";
-             })
-        .def("get_name",
-             [](const ObsDatum<PrecisionT> &obs) { return obs.getObsName(); })
-        .def("get_wires",
-             [](const ObsDatum<PrecisionT> &obs) { return obs.getObsWires(); })
-        .def("get_params", [](const ObsDatum<PrecisionT> &obs) {
-            py::list params;
-            for (size_t i = 0; i < obs.getObsParams().size(); i++) {
-                std::visit(
-                    [&](const auto &param) {
-                        using p_t = std::decay_t<decltype(param)>;
-                        if constexpr (std::is_same_v<
-                                          p_t,
-                                          std::vector<std::complex<ParamT>>>) {
-                            params.append(py::array_t<std::complex<ParamT>>(
-                                py::cast(param)));
-                        } else if constexpr (std::is_same_v<
-                                                 p_t, std::vector<ParamT>>) {
-                            params.append(py::array_t<ParamT>(py::cast(param)));
-                        } else if constexpr (std::is_same_v<p_t,
-                                                            std::monostate>) {
-                            params.append(py::list{});
-                        } else {
-                            throw("Unsupported data type");
-                        }
-                    },
-                    obs.getObsParams()[i]);
-            }
-            return params;
-        });
-
-    //***********************************************************************//
-    //                              Operations
-    //***********************************************************************//
-
-    class_name = "OpsStructC" + bitsize;
-    py::class_<OpsData<PrecisionT>>(m, class_name.c_str(), py::module_local())
-        .def(py::init<
-             const std::vector<std::string> &,
-             const std::vector<std::vector<ParamT>> &,
-             const std::vector<std::vector<size_t>> &,
-             const std::vector<bool> &,
-             const std::vector<std::vector<std::complex<PrecisionT>>> &>())
-        .def("__repr__", [](const OpsData<PrecisionT> &ops) {
-            using namespace Pennylane::Util;
-            std::ostringstream ops_stream;
-            for (size_t op = 0; op < ops.getSize(); op++) {
-                ops_stream << "{'name': " << ops.getOpsName()[op];
-                ops_stream << ", 'params': " << ops.getOpsParams()[op];
-                ops_stream << ", 'inv': " << ops.getOpsInverses()[op];
-                ops_stream << "}";
-                if (op < ops.getSize() - 1) {
-                    ops_stream << ",";
-                }
-            }
-            return "Operations: [" + ops_stream.str() + "]";
-        });
-
-    //***********************************************************************//
-    //                              Adjoint Jacobian
-    //***********************************************************************//
-
-    class_name = "AdjointJacobianC" + bitsize;
-    py::class_<AdjointJacobian<PrecisionT>>(m, class_name.c_str(),
-                                            py::module_local())
-        .def(py::init<>())
-        .def("create_ops_list",
-             [](AdjointJacobian<PrecisionT> &adj,
-                const std::vector<std::string> &ops_name,
-                const std::vector<np_arr_r> &ops_params,
-                const std::vector<std::vector<size_t>> &ops_wires,
-                const std::vector<bool> &ops_inverses,
-                const std::vector<np_arr_c> &ops_matrices) {
-                 std::vector<std::vector<PrecisionT>> conv_params(
-                     ops_params.size());
-                 std::vector<std::vector<std::complex<PrecisionT>>>
-                     conv_matrices(ops_matrices.size());
-                 static_cast<void>(adj);
-                 for (size_t op = 0; op < ops_name.size(); op++) {
-                     const auto p_buffer = ops_params[op].request();
-                     const auto m_buffer = ops_matrices[op].request();
-                     if (p_buffer.size) {
-                         const auto *const p_ptr =
-                             static_cast<const ParamT *>(p_buffer.ptr);
-                         conv_params[op] =
-                             std::vector<ParamT>{p_ptr, p_ptr + p_buffer.size};
-                     }
-                     if (m_buffer.size) {
-                         const auto m_ptr =
-                             static_cast<const std::complex<ParamT> *>(
-                                 m_buffer.ptr);
-                         conv_matrices[op] = std::vector<std::complex<ParamT>>{
-                             m_ptr, m_ptr + m_buffer.size};
-                     }
-                 }
-                 return OpsData<PrecisionT>{ops_name, conv_params, ops_wires,
-                                            ops_inverses, conv_matrices};
-             })
-        .def("adjoint_jacobian", &AdjointJacobian<PrecisionT>::adjointJacobian)
-        .def("adjoint_jacobian",
-             [](AdjointJacobian<PrecisionT> &adj,
-                const StateVectorRawCPU<PrecisionT> &sv,
-                const std::vector<ObsDatum<PrecisionT>> &observables,
-                const OpsData<PrecisionT> &operations,
-                const std::vector<size_t> &trainableParams, size_t num_params) {
-                 std::vector<PrecisionT> jac(observables.size() * num_params,
-                                             0);
-
-                 const JacobianData<PrecisionT> jd{
-                     num_params,  sv.getLength(), sv.getData(),
-                     observables, operations,     trainableParams};
-
-                 adj.adjointJacobian(jac, jd);
-
-                 return py::array_t<ParamT>(py::cast(jac));
-             });
-
-    //***********************************************************************//
-    //                              VJP
-    //***********************************************************************//
-
-    class_name = "VectorJacobianProductC" + bitsize;
-    py::class_<VectorJacobianProduct<PrecisionT>>(m, class_name.c_str(),
-                                                  py::module_local())
-        .def(py::init<>())
-        .def("create_ops_list",
-             [](VectorJacobianProduct<PrecisionT> &v,
-                const std::vector<std::string> &ops_name,
-                const std::vector<np_arr_r> &ops_params,
-                const std::vector<std::vector<size_t>> &ops_wires,
-                const std::vector<bool> &ops_inverses,
-                const std::vector<np_arr_c> &ops_matrices) {
-                 std::vector<std::vector<PrecisionT>> conv_params(
-                     ops_params.size());
-                 std::vector<std::vector<std::complex<PrecisionT>>>
-                     conv_matrices(ops_matrices.size());
-                 static_cast<void>(v);
-                 for (size_t op = 0; op < ops_name.size(); op++) {
-                     const auto p_buffer = ops_params[op].request();
-                     const auto m_buffer = ops_matrices[op].request();
-                     if (p_buffer.size) {
-                         const auto *const p_ptr =
-                             static_cast<const ParamT *>(p_buffer.ptr);
-                         conv_params[op] =
-                             std::vector<ParamT>{p_ptr, p_ptr + p_buffer.size};
-                     }
-                     if (m_buffer.size) {
-                         const auto m_ptr =
-                             static_cast<const std::complex<ParamT> *>(
-                                 m_buffer.ptr);
-                         conv_matrices[op] = std::vector<std::complex<ParamT>>{
-                             m_ptr, m_ptr + m_buffer.size};
-                     }
-                 }
-                 return OpsData<PrecisionT>{ops_name, conv_params, ops_wires,
-                                            ops_inverses, conv_matrices};
-             })
-        .def("compute_vjp_from_jac",
-             &VectorJacobianProduct<PrecisionT>::computeVJP)
-        .def("compute_vjp_from_jac",
-             [](VectorJacobianProduct<PrecisionT> &v,
-                const std::vector<PrecisionT> &jac,
-                const std::vector<PrecisionT> &dy_row, size_t m, size_t n) {
-                 std::vector<PrecisionT> vjp_res(n);
-                 v.computeVJP(vjp_res, jac, dy_row, m, n);
-                 return py::array_t<ParamT>(py::cast(vjp_res));
-             })
-        .def("vjp_fn",
-             [](VectorJacobianProduct<PrecisionT> &v,
-                const std::vector<PrecisionT> &dy, size_t num_params) {
-                 auto fn = v.vectorJacobianProduct(dy, num_params);
-                 return py::cpp_function(
-                     [fn, num_params](
-                         const StateVectorRawCPU<PrecisionT> &sv,
-                         const std::vector<ObsDatum<PrecisionT>> &observables,
-                         const OpsData<PrecisionT> &operations,
-                         const std::vector<size_t> &trainableParams) {
-                         const JacobianData<PrecisionT> jd{
-                             num_params,  sv.getLength(), sv.getData(),
-                             observables, operations,     trainableParams};
-                         return py::array_t<ParamT>(py::cast(fn(jd)));
-                     });
-             });
 
     //***********************************************************************//
     //                              Measures
@@ -371,6 +132,230 @@ void lightning_class_bindings(py::module_ &m) {
         });
 }
 
+template <class PrecisionT, class ParamT>
+void registerAlgorithms(py::module_ &m) {
+    const std::string bitsize =
+        std::to_string(sizeof(std::complex<PrecisionT>) * 8);
+
+    //***********************************************************************//
+    //                              Observable
+    //***********************************************************************//
+
+    using np_arr_c = py::array_t<std::complex<ParamT>, py::array::c_style>;
+    using np_arr_r = py::array_t<ParamT, py::array::c_style>;
+
+    std::string class_name;
+
+    class_name = "ObservableC" + bitsize;
+    py::class_<Observable<PrecisionT>, std::shared_ptr<Observable<PrecisionT>>>(
+        m, class_name.c_str(), py::module_local());
+
+    class_name = "NamedObsC" + bitsize;
+    py::class_<NamedObs<PrecisionT>, std::shared_ptr<NamedObs<PrecisionT>>,
+               Observable<PrecisionT>>(m, class_name.c_str(),
+                                       py::module_local())
+        .def(py::init(
+            [](const std::string &name, const std::vector<size_t> &wires) {
+                return NamedObs<PrecisionT>(name, wires);
+            }))
+        .def("__repr__", &NamedObs<PrecisionT>::getObsName)
+        .def("get_wires", &NamedObs<PrecisionT>::getWires,
+             "Get wires of observables")
+        .def(
+            "__eq__",
+            [](const NamedObs<PrecisionT> &self, py::handle other) -> bool {
+                if (!py::isinstance<NamedObs<PrecisionT>>(other)) {
+                    return false;
+                }
+                auto other_cast = other.cast<NamedObs<PrecisionT>>();
+                return self == other_cast;
+            },
+            "Compare two observables");
+
+    class_name = "HermitianObsC" + bitsize;
+    py::class_<HermitianObs<PrecisionT>,
+               std::shared_ptr<HermitianObs<PrecisionT>>,
+               Observable<PrecisionT>>(m, class_name.c_str(),
+                                       py::module_local())
+        .def(py::init([](const np_arr_c &matrix,
+                         const std::vector<size_t> &wires) {
+            auto buffer = matrix.request();
+            const auto *ptr =
+                static_cast<std::complex<PrecisionT> *>(buffer.ptr);
+            return HermitianObs<PrecisionT>(
+                std::vector<std::complex<PrecisionT>>(ptr, ptr + buffer.size),
+                wires);
+        }))
+        .def("__repr__", &HermitianObs<PrecisionT>::getObsName)
+        .def("get_wires", &HermitianObs<PrecisionT>::getWires,
+             "Get wires of observables")
+        .def(
+            "__eq__",
+            [](const HermitianObs<PrecisionT> &self, py::handle other) -> bool {
+                if (!py::isinstance<HermitianObs<PrecisionT>>(other)) {
+                    return false;
+                }
+                auto other_cast = other.cast<HermitianObs<PrecisionT>>();
+                return self == other_cast;
+            },
+            "Compare two observables");
+
+    class_name = "TensorProdObsC" + bitsize;
+    py::class_<TensorProdObs<PrecisionT>,
+               std::shared_ptr<TensorProdObs<PrecisionT>>,
+               Observable<PrecisionT>>(m, class_name.c_str(),
+                                       py::module_local())
+        .def(py::init(
+            [](const std::vector<std::shared_ptr<Observable<PrecisionT>>>
+                   &obs) { return TensorProdObs<PrecisionT>(obs); }))
+        .def("__repr__", &TensorProdObs<PrecisionT>::getObsName)
+        .def("get_wires", &TensorProdObs<PrecisionT>::getWires,
+             "Get wires of observables")
+        .def(
+            "__eq__",
+            [](const TensorProdObs<PrecisionT> &self,
+               py::handle other) -> bool {
+                if (!py::isinstance<TensorProdObs<PrecisionT>>(other)) {
+                    return false;
+                }
+                auto other_cast = other.cast<TensorProdObs<PrecisionT>>();
+                return self == other_cast;
+            },
+            "Compare two observables");
+
+    class_name = "HamiltonianC" + bitsize;
+    using ObsPtr = std::shared_ptr<Observable<PrecisionT>>;
+    py::class_<Hamiltonian<PrecisionT>,
+               std::shared_ptr<Hamiltonian<PrecisionT>>,
+               Observable<PrecisionT>>(m, class_name.c_str(),
+                                       py::module_local())
+        .def(py::init([](const np_arr_r &coeffs,
+                         const std::vector<ObsPtr> &obs) {
+            auto buffer = coeffs.request();
+            const auto ptr = static_cast<const ParamT *>(buffer.ptr);
+            return Hamiltonian<PrecisionT>{std::vector(ptr, ptr + buffer.size),
+                                           obs};
+        }))
+        .def("__repr__", &Hamiltonian<PrecisionT>::getObsName)
+        .def("get_wires", &Hamiltonian<PrecisionT>::getWires,
+             "Get wires of observables")
+        .def(
+            "__eq__",
+            [](const Hamiltonian<PrecisionT> &self, py::handle other) -> bool {
+                if (!py::isinstance<Hamiltonian<PrecisionT>>(other)) {
+                    return false;
+                }
+                auto other_cast = other.cast<Hamiltonian<PrecisionT>>();
+                return self == other_cast;
+            },
+            "Compare two observables");
+
+    //***********************************************************************//
+    //                              Operations
+    //***********************************************************************//
+
+    class_name = "OpsStructC" + bitsize;
+    py::class_<OpsData<PrecisionT>>(m, class_name.c_str(), py::module_local())
+        .def(py::init<
+             const std::vector<std::string> &,
+             const std::vector<std::vector<ParamT>> &,
+             const std::vector<std::vector<size_t>> &,
+             const std::vector<bool> &,
+             const std::vector<std::vector<std::complex<PrecisionT>>> &>())
+        .def("__repr__", [](const OpsData<PrecisionT> &ops) {
+            using namespace Pennylane::Util;
+            std::ostringstream ops_stream;
+            for (size_t op = 0; op < ops.getSize(); op++) {
+                ops_stream << "{'name': " << ops.getOpsName()[op];
+                ops_stream << ", 'params': " << ops.getOpsParams()[op];
+                ops_stream << ", 'inv': " << ops.getOpsInverses()[op];
+                ops_stream << "}";
+                if (op < ops.getSize() - 1) {
+                    ops_stream << ",";
+                }
+            }
+            return "Operations: [" + ops_stream.str() + "]";
+        });
+
+    /**
+     * Create operation list
+     * */
+    std::string function_name = "create_ops_list_C" + bitsize;
+    m.def(
+        function_name.c_str(),
+        [](const std::vector<std::string> &ops_name,
+           const std::vector<std::vector<PrecisionT>> &ops_params,
+           const std::vector<std::vector<size_t>> &ops_wires,
+           const std::vector<bool> &ops_inverses,
+           const std::vector<np_arr_c> &ops_matrices) {
+            std::vector<std::vector<std::complex<PrecisionT>>> conv_matrices(
+                ops_matrices.size());
+            for (size_t op = 0; op < ops_name.size(); op++) {
+                const auto m_buffer = ops_matrices[op].request();
+                if (m_buffer.size) {
+                    const auto m_ptr =
+                        static_cast<const std::complex<ParamT> *>(m_buffer.ptr);
+                    conv_matrices[op] = std::vector<std::complex<ParamT>>{
+                        m_ptr, m_ptr + m_buffer.size};
+                }
+            }
+            return OpsData<PrecisionT>{ops_name, ops_params, ops_wires,
+                                       ops_inverses, conv_matrices};
+        },
+        "Create a list of operations from data.");
+    m.def(
+        "adjoint_jacobian",
+        [](const StateVectorRawCPU<PrecisionT> &sv,
+           const std::vector<std::shared_ptr<Observable<PrecisionT>>>
+               &observables,
+           const OpsData<PrecisionT> &operations,
+           const std::vector<size_t> &trainableParams) {
+            std::vector<PrecisionT> jac(
+                observables.size() * trainableParams.size(), PrecisionT{0.0});
+
+            const JacobianData<PrecisionT> jd{operations.getTotalNumParams(),
+                                              sv.getLength(),
+                                              sv.getData(),
+                                              observables,
+                                              operations,
+                                              trainableParams};
+
+            adjointJacobian(std::span{jac}, jd);
+
+            return py::array_t<ParamT>(py::cast(jac));
+        },
+        "Compute jacobian of the circuit using the adjoint method.");
+
+    m.def(
+        "statevector_vjp",
+        /* Do not cast non-conforming array. Argument trainableParams should
+         * only contain indices for operations.
+         */
+        [](const StateVectorRawCPU<PrecisionT> &sv,
+           const OpsData<PrecisionT> &operations, const np_arr_c &dy,
+           const std::vector<size_t> &trainableParams) {
+            std::vector<std::complex<PrecisionT>> vjp(
+                trainableParams.size(), std::complex<PrecisionT>{});
+
+            const JacobianData<PrecisionT> jd{operations.getTotalNumParams(),
+                                              sv.getLength(),
+                                              sv.getData(),
+                                              {},
+                                              operations,
+                                              trainableParams};
+            const auto buffer = dy.request();
+
+            statevectorVJP<PrecisionT>(
+                std::span{vjp}, jd,
+                std::span{
+                    static_cast<const std::complex<PrecisionT> *>(buffer.ptr),
+                    static_cast<size_t>(buffer.size)});
+
+            return py::array_t<std::complex<PrecisionT>>(py::cast(vjp));
+        },
+        "Compute jacobian of the circuit using the adjoint method.");
+}
+
 /**
  * @brief Add C++ classes, methods and functions to Python module.
  */
@@ -404,6 +389,13 @@ PYBIND11_MODULE(lightning_qubit_ops, // NOLINT: No control over Pybind internals
           py::overload_cast<const vector<size_t> &, size_t>(
               &Gates::getIndicesAfterExclusion),
           "Get statevector indices for gate application");
+
+    /* Algorithms submodule */
+    py::module_ alg_submodule = m.def_submodule(
+        "adjoint_diff", "A submodule for adjoint differentiation method.");
+
+    registerAlgorithms<float, float>(alg_submodule);
+    registerAlgorithms<double, double>(alg_submodule);
 
     /* Add CPUMemoryModel enum class */
     py::enum_<CPUMemoryModel>(m, "CPUMemoryModel")