+++ pennylane_lightning/src/util/Memory.hpp
@@ -67,6 +67,9 @@ inline void alignedFree(void *p) {
 /**
  * @brief C++ Allocator class for aligned memory.
  *
+ * C++17's std::pmr::polymorphic_allocator substitutes this whole class.
+ * However, clang (even the latest version 13) does not support pmr yet.
+ *
  * @tparam T Datatype to allocate
  */
 template <class T> class AlignedAllocator {
@@ -107,7 +110,7 @@ template <class T> class AlignedAllocator {
      * @param size The number of T objects for the allocation
      * @return Allocated aligned memory
      */
-    [[nodiscard]] T *allocate(std::size_t size) {
+    [[nodiscard]] T *allocate(std::size_t size) const {
         if (size == 0) {
             return nullptr;
         }
@@ -130,7 +133,7 @@ template <class T> class AlignedAllocator {
      * @param p Pointer to the allocated data
      * @param size Size of the data we allocated (unused).
      */
-    void deallocate(T *p, [[maybe_unused]] std::size_t size) noexcept {
+    void deallocate(T *p, [[maybe_unused]] std::size_t size) const noexcept {
         if (alignment_ > alignof(std::max_align_t)) {
             alignedFree(p);
         } else {